+++ src/1_resources.js
@@ -175,7 +175,6 @@ resources.SMSLinkSend = {
 	params: {
 		"phone": validator(true, validationTypes.str)
 	}
-
 };
 
 resources.event = {